Split types into cinema and movie type files
